requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-default-group==1.2.4
condense-json==0.1.2
defusedxml==0.7.1
distro==1.9.0
faster-whisper==1.0.3
ffmpeg-python==0.2.0
filelock==3.18.0
Flask==3.1.0
flask-cors==5.0.1
fsspec==2025.3.2
future==1.0.0
google-auth==2.38.0
google-genai==1.8.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.30.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
llm==0.24.2
MarkupSafe==3.0.2
more-itertools==10.7.0
ollama==0.4.8
packaging==24.2
pillow==11.2.1
pluggy==1.5.0
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.1
pydantic_core==2.33.0
python-dateutil==2.9.0.post0
python-ulid==3.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.3
scipy==1.15.2
semantic-version==2.10.0
setuptools==80.3.0
setuptools-rust==1.11.1
six==1.17.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-migrate==0.1b0
sqlite-utils==3.38
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0+cpu
torchaudio==2.7.0+cpu
torchvision==0.22.0+cpu
tqdm==4.67.1
transformers==4.50.3
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uuid==1.30
waitress==3.0.2
websockets==15.0.1
Werkzeug==3.1.3
youtube-transcript-api==1.0.3
